fix: short to long

+++ Blockchain/Laboratorios/Laboratorios-resueltos/TP-final/apps/web/src/wallet/WalletControls.tsx
@@ -81,7 +81,7 @@ function WalletControlsContent() {
           {switchError && <span className="text-xs text-danger">No se cambió la red.</span>}
         </div>
       )}
-      <AddressDisplay address={address!} format="short" tooltipSide="bottom" />
+      <AddressDisplay address={address!} format="long" tooltipSide="bottom" />
       <Button variant="ghost" onClick={() => disconnect()}>
         Desconectar
       </Button>
